+++ packages/astro/test/fixtures/before-hydration/astro.config.mjs
@@ -1,0 +1,6 @@
+import { defineConfig } from 'astro/config';
+import preact from '@astrojs/preact';
+
+export default defineConfig({
+	integrations: [preact()]
+});